+++ tests/testthat/test-citations.R
@@ -14,7 +14,7 @@ test_that("documents with spaces in names can be rendered", {
 
   write(c("", "Optional reference: @@reference2020"),
         file = file_name, append = TRUE)
-  if (rmarkdown::pandoc_available("1.14")){
+  if (rmarkdown::pandoc_available("2.0")){
     sink_out <- capture_output(cite_all(file_name))
 
     contents <- readLines(gsub("Rmd$", "md", file_name), encoding = "UTF-8")
@@ -48,7 +48,7 @@ test_that("cite_* work with umlaute", {
   #      file = file_name, append = TRUE)
   worcs:::write_as_utf(c("", "Optional räference: @@reference2020"),
                        con = file_name, append = TRUE)
-  if (rmarkdown::pandoc_available("1.14")){
+  if (rmarkdown::pandoc_available("2.0")){
     sink_out <- capture_output(cite_all(file_name))
     md_file <- force(gsub("Rmd$", "md", file_name))
     contents <- worcs:::read_as_utf(md_file)
@@ -91,7 +91,7 @@ myiris <- read.csv('iris.csv')
 "), file = file_name, append = TRUE)
 
   # Render with all citations
-  if (rmarkdown::pandoc_available("1.14")){
+  if (rmarkdown::pandoc_available("2.0")){
     sink_out <- capture_output(cite_all(file_name))
 
     contents <- readLines(normalizePath(gsub("\\.Rmd", "\\.md", file_name)))
@@ -124,7 +124,7 @@ test_that("cite_* retain double @", {
 
   write(c("", "Optional reference: @@reference2020"),
         file = file_name, append = TRUE)
-  if (rmarkdown::pandoc_available("1.14")){
+  if (rmarkdown::pandoc_available("2.0")){
     sink_out <- capture_output(cite_all(file_name))
 
     contents <- readLines(file_name, encoding = "UTF-8")